backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
mysql-connector-python
SQLAlchemy
pytest
httpx

